# specs/001-purple-glow-social/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Purple Glow Social API
  version: 1.0.0
  description: |
    RESTful API for Purple Glow Social - AI-powered social media management platform.

    **Authentication**: All endpoints (except /auth/*) require Bearer token in Authorization header.

    **Rate Limiting**: 
    - Authenticated: 100 req/min per user
    - Unauthenticated: 10 req/min per IP

    **Error Responses**: All endpoints return RFC 7807 Problem Details for HTTP APIs format.
  contact:
    name: Purple Glow Social Support
    email: support@purpleglowsocial.com
  license:
    name: Proprietary
servers:
  - url: https://api.purpleglowsocial.com/v1
    description: Production server
  - url: http://localhost:3000/api/v1
    description: Local development server
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token from Supabase Auth
  schemas:
    Error:
      type: object
      required:
        - type
        - title
        - status
      properties:
        type:
          type: string
          format: uri
          example: 'https://api.purpleglowsocial.com/errors/validation-error'
        title:
          type: string
          example: 'Validation Error'
        status:
          type: integer
          example: 400
        detail:
          type: string
          example: 'The request body contains invalid data.'
        instance:
          type: string
          format: uri
          example: '/api/v1/posts'
        errors:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
...
    Subscription:
      type: object
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        tier:
          type: string
          enum: [starter, growth, enterprise]
        status:
          type: string
          enum: [trial, active, past_due, cancelled, expired]
        billing_cycle_start:
          type: string
          format: date
        next_billing_date:
          type: string
          format: date
        trial_end_date:
          type: string
          format: date
          nullable: true
        posts_used_current_cycle:
          type: integer
        posts_limit:
          type: integer
        platforms_connected:
          type: integer
        platforms_limit:
          type: integer
        team_members_count:
          type: integer
        team_members_limit:
          type: integer
        storage_used_mb:
          type: number
          format: float
        storage_limit_mb:
          type: integer
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
  responses:
    UnauthorizedError:
      description: Authentication token is missing or invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            type: 'https://api.purpleglowsocial.com/errors/unauthorized'
            title: 'Unauthorized'
            status: 401
            detail: 'Invalid or expired authentication token.'
...
security:
  - BearerAuth: []
paths:
...
  /subscriptions/me:
    get:
      tags:
        - Subscriptions
      summary: Get current subscription
      responses:
        '200':
          description: Subscription retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Subscription'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /subscriptions/upgrade:
    post:
      tags:
        - Subscriptions
      summary: Upgrade subscription tier
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - tier
              properties:
                tier:
                  type: string
                  enum: [growth, enterprise]
      responses:
        '200':
          description: Upgrade initiated (Paystack payment URL returned)
          content:
            application/json:
              schema:
                type: object
                properties:
                  payment_url:
                    type: string
                    format: uri
                  reference:
                    type: string
        '400':
          description: Invalid tier or downgrade not allowed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /subscriptions/cancel:
    post:
      tags:
        - Subscriptions
      summary: Cancel subscription
      description: Cancel recurring billing (access remains until end of cycle)
      responses:
        '200':
          description: Subscription cancelled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Subscription'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
...